+++ polaris-react/src/components/FullscreenBar/FullscreenBar.stories.tsx
@@ -53,8 +53,8 @@ export function WithChildren() {
   const titleMarkup = (
     <div
       style={{
-        marginLeft: 'var(--p-space-2)',
-        marginRight: 'var(--p-space-4)',
+        marginLeft: 'var(--p-space-200)',
+        marginRight: 'var(--p-space-400)',
         flexGrow: 1,
       }}
     >