+++ polaris-react/src/components/SkeletonPage/SkeletonPage.tsx
@@ -70,8 +70,8 @@ export function SkeletonPage({
       <Box
         width="100%"
         padding="0"
-        paddingInlineStart={{sm: '6'}}
-        paddingInlineEnd={{sm: '6'}}
+        paddingInlineStart={{sm: '600'}}
+        paddingInlineEnd={{sm: '600'}}
         maxWidth="var(--pc-skeleton-page-max-width)"
         aria-label={i18n.translate('Polaris.SkeletonPage.loadingLabel')}
         role="status"
@@ -84,23 +84,23 @@ export function SkeletonPage({
       >
         <BlockStack>
           <Box
-            paddingBlockStart={{xs: '4', md: '5'}}
-            paddingBlockEnd={{xs: '4', md: '5'}}
-            paddingInlineStart={{xs: '4', sm: '0'}}
-            paddingInlineEnd={{xs: '4', sm: '0'}}
+            paddingBlockStart={{xs: '400', md: '500'}}
+            paddingBlockEnd={{xs: '400', md: '500'}}
+            paddingInlineStart={{xs: '400', sm: '0'}}
+            paddingInlineEnd={{xs: '400', sm: '0'}}
             width="100%"
           >
             <InlineStack gap="4" align="space-between" blockAlign="center">
               <InlineStack gap="4">
                 {backActionMarkup}
-                <Box paddingBlockStart="1" paddingBlockEnd="1">
+                <Box paddingBlockStart="100" paddingBlockEnd="100">
                   {titleContent}
                 </Box>
               </InlineStack>
               {primaryActionMarkup}
             </InlineStack>
           </Box>
-          <Box paddingBlockEnd="2" width="100%">
+          <Box paddingBlockEnd="200" width="100%">
             {children}
           </Box>
         </BlockStack>