+++ polaris.shopify.com/CHANGELOG.md
@@ -1,5 +1,27 @@
 # polaris.shopify.com
 
+## 0.58.0
+
+### Minor Changes
+
+- [#10585](https://github.com/Shopify/polaris/pull/10585) [`45fc3a01f`](https://github.com/Shopify/polaris/commit/45fc3a01f959f6eea2c1bf03b981811a6f13f344) Thanks [@laurkim](https://github.com/laurkim)! - Migrated `space` custom properties from `v11` to `v12`
+
+### Patch Changes
+
+- Updated dependencies [[`45fc3a01f`](https://github.com/Shopify/polaris/commit/45fc3a01f959f6eea2c1bf03b981811a6f13f344)]:
+  - @shopify/polaris@11.21.0
+  - @shopify/polaris-tokens@7.12.0
+
+## 0.57.9
+
+### Patch Changes
+
+- [#10609](https://github.com/Shopify/polaris/pull/10609) [`b53189c72`](https://github.com/Shopify/polaris/commit/b53189c727bd62d67b641ae906b3af1a5bf60de7) Thanks [@kyledurand](https://github.com/kyledurand)! - Updated component example backgrounds
+
+- Updated dependencies [[`22dbada34`](https://github.com/Shopify/polaris/commit/22dbada3490b8ed5c5c65c77c532529b411128e2), [`f26cebe99`](https://github.com/Shopify/polaris/commit/f26cebe99e486730bce2ca1d9add89e24ef9c648), [`f6a965102`](https://github.com/Shopify/polaris/commit/f6a965102f71820375e764e6bdbb431fb9542273), [`3bafbea50`](https://github.com/Shopify/polaris/commit/3bafbea5000207c05d2c58dc3553790164ab7844), [`2359e5f5a`](https://github.com/Shopify/polaris/commit/2359e5f5a4e6a93f8dd70997c4ae9a1ab732911c)]:
+  - @shopify/polaris@11.20.1
+  - @shopify/polaris-tokens@7.11.0
+
 ## 0.57.8
 
 ### Patch Changes